moved all functions originating from Browse parent component into browse component, helps with calling re renders getting page properly updated
