+++ client/src/components/BrowseContainer/index.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import axios from 'axios';
 import Col from 'react-bootstrap/Col';
 import Row from "react-bootstrap/Row"
 import Button from "react-bootstrap/Button";
