+++ client/src/pages/Browse.js
@@ -12,7 +12,18 @@ function Browse() {
     const [itemArray, setItemArray] = useState([]);
     //useEffect loads once when page renders calling async fetchData
     const { userId } = useParams()
+    //declare variables for ItemnModal
+    const [newItemInfo, setNewItemInfo] = useState({
+        name: '',
+        description: '',
+        price: 0,
+        imgUrl: "",
+    });
+    // declare setState variables to show or hide itemModal
+    const [show, setShow] = useState(false);
     
+    
+
     useEffect(() => {
 
         if (userId) {
@@ -43,6 +54,36 @@ function Browse() {
         // Calls useEffect anytime itemArray is changed
     }, []);
 
+    //Show or hide newItemModal
+    const handleShow = () => setShow(true);
+    const closeModal = () => setShow(false);
+
+    // Handle submit newItemModal
+    const handleSubmit = (event) => {
+        event.preventDefault();
+        setShow(false);
+        axios
+            .put('/api/users/addItem', {
+                item: {
+                    name: newItemInfo.name,
+                    description: newItemInfo.description,
+                    price: newItemInfo.price,
+                    imgUrl: newItemInfo.imgUrl,
+                }
+            })
+            .then((response) => {
+                setItemArray(response.data.items)
+            })
+            .catch((error) => {
+                console.log(error);
+            });
+    };
+
+    const handleChange = (event) => {
+        const { name, value } = event.target;
+        setNewItemInfo({ ...newItemInfo, [name]: value });
+    };
+
     const handleDelete = (_id) => {
         axios
             .put('/api/users/delItem', {
@@ -72,6 +113,14 @@ function Browse() {
                     <Row className="d-inline-flex">
                         <Col className="col" >
                             <ItemModal
+                                handleShow={handleShow}
+                                closeModal={closeModal}
+                                show={show}
+                                setShow={setShow}
+                                newItemInfo={newItemInfo}
+                                setNewItemInfo={setNewItemInfo}
+                                handleChange={handleChange}
+                                handleSubmit={handleSubmit}
                                 setItemArray={setItemArray}
                                 itemArray={itemArray}
                             />
